feat: make the button one of the bottom bar
