+++ packages/cad-viewer/src/component/statusBar/index.ts
@@ -1,1 +1,2 @@
 export { default as MlStatusBar } from './MlStatusBar.vue'
+export { default as MlExportButton } from './MlExportButton.vue'
